fix(workflows): re-organize files
