+++ workflows/quickstart.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package execute
+package workflows
 
 // [START workflows_api_quickstart]
 import (
